# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Identrail API
  version: "1.0.0"
  description: |
    V1 machine identity security API for AWS and Kubernetes discovery, graphing, and findings.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /auth/logout:
    post:
      tags: [Authorization]
      summary: Log out the current session
      operationId: logoutCurrentSession
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Current session revoked and cookie cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/login:
    get:
      tags: [Authorization]
      summary: Start WorkOS AuthKit sign-in
      operationId: startWorkOSLogin
      security: []
      parameters:
        - in: query
          name: return_to
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to WorkOS AuthKit
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/signup:
    get:
      tags: [Authorization]
      summary: Start WorkOS AuthKit sign-up
      operationId: startWorkOSSignup
      security: []
      parameters:
        - in: query
          name: return_to
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to WorkOS AuthKit sign-up screen
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/callback:
    get:
      tags: [Authorization]
      summary: Complete WorkOS AuthKit callback
      operationId: completeWorkOSCallback
      security: []
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Session cookie set and browser redirected to the app
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/webhooks/workos:
    post:
      tags: [Authorization]
      summary: Receive WorkOS AuthKit lifecycle webhooks
      operationId: receiveWorkOSWebhook
      security:
        - WorkOSWebhookSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/auth/config:
    parameters:
      - $ref: "#/components/parameters/scopeTenantID"
      - $ref: "#/components/parameters/scopeWorkspaceID"
    get:
      tags: [Authorization]
      summary: Get public authentication configuration
      operationId: getAuthConfig
      security: []
      responses:
        "200":
          description: Authentication options available to the frontend
          content:
            application/json:
              schema:
                type: object
                required: [auth]
                properties:
                  auth:
                    type: object
                    required: [manual_mode, workos_login_enabled, providers]
                    properties:
                      manual_mode:
                        type: boolean
                      workos_login_enabled:
                        type: boolean
                      providers:
                        type: array
                        items:
                          type: string
  /v1/me:
    parameters:
      - $ref: "#/components/parameters/scopeTenantID"
      - $ref: "#/components/parameters/scopeWorkspaceID"
    get:
      tags: [Authorization]
      summary: Get current session user
      operationId: getCurrentUser
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Current user and session context
          content:
            application/json:
              schema:
                type: object
                properties:
                  me:
                    $ref: "#/components/schemas/CurrentUserContext"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/sessions:
    parameters:
      - $ref: "#/components/parameters/scopeTenantID"
      - $ref: "#/components/parameters/scopeWorkspaceID"
    get:
      tags: [Authorization]
      summary: List current user sessions
      operationId: listCurrentUserSessions
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Active sessions for the current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SessionListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/sessions/{session_id}:
    parameters:
      - $ref: "#/components/parameters/scopeTenantID"
      - $ref: "#/components/parameters/scopeWorkspaceID"
      - in: path
        name: session_id
        required: true
        schema: { type: string }
    delete:
      tags: [Authorization]
      summary: Revoke one current-user session
      operationId: revokeCurrentUserSession
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/me/sessions/revoke-others:
    parameters:
      - $ref: "#/components/parameters/scopeTenantID"
      - $ref: "#/components/parameters/scopeWorkspaceID"
    post:
      tags: [Authorization]
      summary: Revoke all other current-user sessions
      operationId: revokeOtherCurrentUserSessions
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Other sessions revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: identrail_session
    GitHubWebhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: GitHub HMAC SHA-256 signature of the raw webhook request body.
    WorkOSWebhookSignature:
      type: apiKey
      in: header
      name: WorkOS-Signature
      description: WorkOS HMAC SHA-256 signature of the raw webhook request body.
  parameters:
    scopeTenantID:
      in: header
      name: X-Identrail-Tenant-ID
      required: false
      schema:
        type: string
      description: Optional tenant scope override when token claims do not provide tenant context. API key callers must use a key bound to the same tenant scope.
    scopeWorkspaceID:
      in: header
      name: X-Identrail-Workspace-ID
      required: false
      schema:
        type: string
      description: Optional workspace scope override when token claims do not provide workspace context. API key callers must use a key bound to the same workspace scope.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        primary_email: { type: string, format: email }
        display_name: { type: string }
        avatar_url: { type: string }
        status:
          type: string
          enum: [active, deactivated, deleted]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
    CurrentUserContext:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        org_id: { type: string }
        workspace_id: { type: string }
        project_id: { type: string }
        role:
          type: string
          enum: [owner, admin, analyst, viewer]
        organization:
          $ref: "#/components/schemas/Organization"
        workspace:
          $ref: "#/components/schemas/Workspace"
        project:
          $ref: "#/components/schemas/Project"
    SessionListItem:
      type: object
      properties:
        id: { type: string }
        ip: { type: string }
        user_agent: { type: string }
        auth_method:
          type: string
          enum: [workos, oidc, manual]
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        idle_expires_at:
          type: string
          format: date-time
        current:
          type: boolean
    Organization:
      type: object
      properties:
        tenant_id: { type: string }
        display_name: { type: string }
        slug: { type: string }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Workspace:
      type: object
      properties:
        tenant_id: { type: string }
        workspace_id: { type: string }
        display_name: { type: string }
        slug: { type: string }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Project:
      type: object
      properties:
        tenant_id: { type: string }
        workspace_id: { type: string }
        project_id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        archived_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...